refactor(apphost): extract WithClearDatabaseCommand into MongoDbResourceBuilderExtensions (#262)
## Summary

Extracts the inline `WithCommand` clear-data block from `AppHost.cs`
into a new `MongoDbResourceBuilderExtensions` class.

## Changes

- **New**: `src/AppHost/MongoDbResourceBuilderExtensions.cs` — contains
`WithMongoDbDevCommands` public entry point and private
`WithClearDatabaseCommand`
- **Simplified**: `src/AppHost/AppHost.cs` — reduced from ~157 lines to
~30 lines; single `mongo.WithMongoDbDevCommands("myblog")` call

## Testing

All 10 existing tests pass:
- 5 unit tests in `MongoDbClearCommandTests`
- 5 integration tests in `MongoClearDataIntegrationTests`

Closes #259
Working as Sam (Backend/.NET)
